Added start/stop file numbers for dark/flats
