+++ pines_analysis_toolkit/data/dome_flat_field.py
@@ -23,11 +23,10 @@ from datetime import datetime
 	Inputs:
 		date (str): the UT date during which the dome flat field data was obtained (i.e., '20200531')
         band (str): the band in which the flat field data was takend, 'J' or 'H'
-        lights_on_start (int, optional): The file number that represents the start of the lights on dome flat sequence. Can use these arguments
+        flat_start (int, optional): The file number that represents the start of the lights **on** dome flat sequence. Can use these arguments
             to specify the lights on/off data, in case the appropriate comments weren't entered into the FITS headers when they were created.
-        lights_on_end (int, optional): The file number that represents the end of the lights on dome flat sequence.
-        lights_off_start (int, optional): The file number that represents the start of the lights off dome flat sequence.
-        lights_off_end (int, optional): The file number that represents the end of the lights off dome flat sequence.
+        flat_end (int, optional): The file number that represents the end of the lights **off** dome flat sequence.
+            NOTE: This will only work if the lights_on sequence and lights_off sequence were taken back-to-back (which should always be the case).
         upload (bool, optional): Whether or not to upload the master dome flat to pines.bu.edu. By default, set to False (so you will not attempt to upload!).
         delete_raw (bool, optional): Whether or not to delete raw data from your local machine when the master flat is created. By default, set to False (won't delete raw dome flat images by default!).
     Outputs:
@@ -35,8 +34,6 @@ from datetime import datetime
 	TODO:
 		Tabular printing
         Bad image flagging
-        Implement specified start/stop arguments
-
 '''
 
 
@@ -77,21 +74,66 @@ def dome_flat_field(date, band, lights_on_start=0, lights_on_stop=0, lights_off_
         sftp.chdir('..')
     
     if data_path == '':
-        print('ERROR: specified date not found in any run on pines.bu.edu:data/raw/mimir/')
+        print('ERROR: {} not found in any run on pines.bu.edu:data/raw/mimir/.'.format(date))
         return
     
     else:
-        log_path = pines_path/'Logs'
-        #Check if you already have the log for this date, if not, download it. 
-        if not (log_path/(date+'_log.txt')).exists():
-            print('Downloading {}_log.txt to {}'.format(date,log_path))
-            sftp.get(date+'_log.txt',log_path/(date+'_log.txt'))
+        #If the file start/stop numbers are specfied, grab those files.
+        if (lights_on_stop != 0) or (lights_off_stop != 0):
+            files_in_dir = sftp.listdir()
+            on_flat_filenums = np.arange(lights_on_start, lights_on_stop+1, step=1)
+            off_flat_filenums = np.arange(lights_off_start, lights_off_stop+1, step=1)
+            flat_files = []
+            lights_on_files = []
+            lights_off_files = []
 
-        log = pines_log_reader(log_path/(date+'_log.txt'))
+            #Add the lights-on flats to the file list. 
+            for i in range(len(on_flat_filenums)):
+                file_num = on_flat_filenums[i]
+                #Generate the filename. 
+                if file_num < 10:
+                    file_name = date+'.00'+str(file_num)+'.fits'
+                elif (file_num >= 10) and (file_num < 100):
+                    file_name = date+'.0'+str(file_num)+'.fits'
+                else:
+                    file_name = date+'.'+str(file_num)+'.fits'
+                #Check if the file name is in the directory, and if so, append it to the list of flat files. 
+                if file_name in files_in_dir:
+                    flat_files.append(file_name)
+                    lights_on_files.append(file_name)
+                else:
+                    print('{} not found in directory, skipping.'.format(file_name))        
 
-        #Identify flat files. 
-        flat_inds = np.where((log['Target'] == 'Flat') & (log['Filename'] != 'test.fits') & (log['Filt.'] == band))[0]
-        flat_files = natsort.natsorted(list(set(log['Filename'][flat_inds]))) #Set guarantees we only grab the unique files that have been identified as flats, in case the log bugged out. 
+            #Do the same for the lights-off files. 
+            for i in range(len(off_flat_filenums)):
+                file_num = off_flat_filenums[i]
+                #Generate the filename. 
+                if file_num < 10:
+                    file_name = date+'.00'+str(file_num)+'.fits'
+                elif (file_num >= 10) and (file_num < 100):
+                    file_name = date+'.0'+str(file_num)+'.fits'
+                else:
+                    file_name = date+'.'+str(file_num)+'.fits'
+                #Check if the file name is in the directory, and if so, append it to the list of flat files. 
+                if file_name in files_in_dir:
+                    flat_files.append(file_name)
+                    lights_off_files.append(file_name)
+                else:
+                    print('{} not found in directory, skipping.'.format(file_name)) 
+        #Otherwise, find the files automatically using the night's log. 
+        else:
+            log_path = pines_path/'Logs'
+            #Check if you already have the log for this date, if not, download it. 
+            if not (log_path/(date+'_log.txt')).exists():
+                print('Downloading {}_log.txt to {}'.format(date,log_path))
+                sftp.get(date+'_log.txt',log_path/(date+'_log.txt'))
+
+            log = pines_log_reader(log_path/(date+'_log.txt'))
+            
+            #Identify flat files. 
+            flat_inds = np.where((log['Target'] == 'Flat') & (log['Filename'] != 'test.fits') & (log['Filt.'] == band))[0]
+            flat_files = natsort.natsorted(list(set(log['Filename'][flat_inds]))) #Set guarantees we only grab the unique files that have been identified as flats, in case the log bugged out. 
+
         print('Found {} flat files.'.format(len(flat_files)))
         print('')
 
@@ -105,26 +147,30 @@ def dome_flat_field(date, band, lights_on_start=0, lights_on_stop=0, lights_off_
                 print('{} already in {}, skipping download.'.format(flat_files[j],dome_flat_raw_path))
         print('')
 
-        #Find the lights-on and lights-off flat files. 
-        lights_on_files = []
-        lights_off_files = []
-        for j in range(len(flat_files)):
-            header = fits.open(dome_flat_raw_path/flat_files[j])[0].header
-            if header['FILTNME2'] != band:
-                print('ERROR: {} taken in filter other than {}. Double check your date, try specifying start/stop file numbers, etc.'.format(flat_files[j], band))
-                return
-            if header['OBJECT'] == 'dome_lamp_on':
-                lights_on_files.append(flat_files[j])
-            elif header['OBJECT'] == 'dome_lamp_off':
-                lights_off_files.append(flat_files[j])
-            else:
-                print("ERROR: header['OBJECT'] for {} is not 'dome_lamp_on' or 'dome_lamp_off. Double check your date, try specifying start/stop file numbers, etc.".format(flat_files[j]))
+        if (lights_on_stop == 0) and (lights_off_stop == 0):
+            #Find the lights-on and lights-off flat files. 
+            lights_on_files = []
+            lights_off_files = []
+            for j in range(len(flat_files)):
+                header = fits.open(dome_flat_raw_path/flat_files[j])[0].header
+                if header['FILTNME2'] != band:
+                    print('ERROR: {} taken in filter other than {}. Double check your date, try specifying start/stop file numbers, etc.'.format(flat_files[j], band))
+                    return
+                if header['OBJECT'] == 'dome_lamp_on':
+                    lights_on_files.append(flat_files[j])
+                elif header['OBJECT'] == 'dome_lamp_off':
+                    lights_off_files.append(flat_files[j])
+                else:
+                    print("ERROR: header['OBJECT'] for {} is not 'dome_lamp_on' or 'dome_lamp_off. Double check your date, try specifying start/stop file numbers, etc.".format(flat_files[j]))
 
         print('Found {} lights-on flat files.'.format(len(lights_on_files)))
         print('Found {} lights-off flat files.'.format(len(lights_off_files)))
         print('')
         time.sleep(1)
 
+        if (len(lights_on_files) == 0) or (len(lights_off_files) == 0):
+            print('ERROR: No lights on/off files fount. Returning.')
+            return
         #Make cube of the lights-on images.
         num_images = len(lights_on_files)
         print('Reading in ', num_images,' lights-on flat images.')
@@ -136,6 +182,10 @@ def dome_flat_field(date, band, lights_on_start=0, lights_on_stop=0, lights_off_
         print('-------------------------------------------------')
         for j in range(len(lights_on_files)):
             image_data = fits.open(dome_flat_raw_path/lights_on_files[j])[0].data[0:1024,:]
+            header = fits.open(dome_flat_raw_path/lights_on_files[j])[0].header
+            if header['FILTNME2'] != band:
+                print('ERROR: {} taken in filter other than {}. Double check your date, try specifying start/stop file numbers, etc.'.format(lights_on_files[j], band))
+                return
             flat_lights_on_cube_raw[j,:,:] = image_data #This line trims off the top two rows of the image, which are overscan.
             print(str(j+1)+'    '+str(np.mean(image_data))+'    '+str(np.std(image_data))+'    '+ str(np.std(image_data))+'    '+str(np.amin(image_data)))
         time.sleep(1)
@@ -151,6 +201,10 @@ def dome_flat_field(date, band, lights_on_start=0, lights_on_stop=0, lights_off_
         print('-------------------------------------------------')
         for j in range(len(lights_off_files)):
             image_data = fits.open(dome_flat_raw_path/lights_off_files[j])[0].data[0:1024,:]
+            header = fits.open(dome_flat_raw_path/lights_off_files[j])[0].header
+            if header['FILTNME2'] != band:
+                print('ERROR: {} taken in filter other than {}. Double check your date, try specifying start/stop file numbers, etc.'.format(lights_off_files[j], band))
+                return
             flat_lights_off_cube_raw[j,:,:] = image_data #This line trims off the top two rows of the image, which are overscan.
             print(str(j+1)+'    '+str(np.mean(image_data))+'    '+str(np.std(image_data))+'    '+ str(np.std(image_data))+'    '+str(np.amin(image_data)))
         time.sleep(1)
