+++ tests/flat_creator_tester.py
@@ -1,5 +1,5 @@
 import pines_analysis_toolkit as pat
 
-run = '20200307'
+date = '20200512'
 band = 'H'
-pat.data.dome_flat_field(run, band, delete_raw=True, upload=True)
+pat.data.dome_flat_field(date, band, lights_on_start=1, lights_on_stop=100, lights_off_start=101, lights_off_stop=200, upload=True, delete_raw=True)
