--- tests/reduce_test.py
@@ -1,5 +1,0 @@
-import pines_analysis_toolkit as pat
-
-target = '2MASS J16291840+0335371'
-
-pat.reduction.reduce(target, upload='True')
